Fix compiler warnings

+++ Visualizer/src/Mesh.cpp
@@ -6,7 +6,7 @@ Mesh::Mesh(const std::vector<Vertex>& vertices, const std::vector<unsigned int>&
     this->textures = textures;
 
 	vertexArray = std::make_unique<VertexArray>();
-	vertexBuffer = std::make_unique<VertexBuffer>(vertices.data(), vertices.size() * sizeof(Vertex));
+	vertexBuffer = std::make_unique<VertexBuffer>(vertices.data(), static_cast<unsigned int>(vertices.size() * sizeof(Vertex)));
 	VertexBufferLayout bufferLayout;
 	bufferLayout.Push<float>(3);
 	bufferLayout.Push<float>(3);
@@ -14,5 +14,5 @@ Mesh::Mesh(const std::vector<Vertex>& vertices, const std::vector<unsigned int>&
 	bufferLayout.Push<float>(2);
 	vertexArray->AddBuffer(*vertexBuffer, bufferLayout);
 
-	indexBuffer = std::make_unique<IndexBuffer>(indices.data(), indices.size());
+	indexBuffer = std::make_unique<IndexBuffer>(indices.data(), static_cast<unsigned int>(indices.size()));
 }
