+++ Visualizer/src/tests/Test3DCube.cpp
@@ -51,12 +51,12 @@ namespace Test {
 		renderer.EnableDepth();
 
 		vertexArray = std::make_unique<VertexArray>();
-		vertexBuffer = std::make_unique<VertexBuffer>(positions, sizeof(positions));
+		vertexBuffer = std::make_unique<VertexBuffer>(positions, static_cast<unsigned int>(sizeof(positions)));
 		VertexBufferLayout bufferLayout;
 		bufferLayout.Push<float>(3);
 		vertexArray->AddBuffer(*vertexBuffer, bufferLayout);
 
-		indexBuffer = std::make_unique<IndexBuffer>(indices, lengthof(indices));
+		indexBuffer = std::make_unique<IndexBuffer>(indices, static_cast<unsigned int>(lengthof(indices)));
 
 		shader = std::make_unique<Shader>("resources/shaders/ClearColor.glsl");
 	}
